requirements.txt
torch_geometric==2.2
yacs 
torchmetrics
performer-pytorch
ogb
wandb
torch_sparse
torch_scatter
fire